requirements.txt
langgraph
langchain-core
openai
requests
python-dotenv
sentence-transformers
torch
numpy
fastapi
uvicorn
httpx

# Voice Gateway (STT + TTS)
deepgram-sdk
sounddevice
soundfile

